move composables to WearApp.kt
